@@ -4,16 +4,15 @@ plugins {
     id("com.github.ben-manes.versions") version("0.51.0")
 }
 
-// Change to true when releasing
-val release = false
 val majorVersion = "1.14.1"
-val minorVersion = if (release) "Release" else "DEV-" + System.getenv("BUILD_NUMBER")
+val minorVersion = "DEV-dxnnv"
 
 group = "com.extendedclip"
 version = "$majorVersion-$minorVersion"
 
 repositories {
     mavenCentral()
+    maven("https://repo.papermc.io/repository/maven-public/")
     maven("https://s01.oss.sonatype.org/content/repositories/snapshots/")
     maven("https://repo.extendedclip.com/content/repositories/placeholderapi/")
     maven("https://repo.glaremasters.me/repository/public/")
@@ -23,7 +22,7 @@ repositories {
 }
 
 dependencies {
-    compileOnly(libs.spigot)
+    compileOnly(libs.paper)
 
     compileOnly(libs.vault)
     compileOnly(libs.authlib)
@@ -39,8 +38,6 @@ dependencies {
     compileOnly(libs.papi)
 
     implementation(libs.nashorn)
-    implementation(libs.adventure.platform)
-    implementation(libs.adventure.minimessage)
 
     compileOnly("org.jetbrains:annotations:23.0.0")
 }
@@ -49,12 +46,11 @@ tasks {
     shadowJar {
         relocate("org.objectweb.asm", "com.extendedclip.deluxemenus.libs.asm")
         relocate("org.openjdk.nashorn", "com.extendedclip.deluxemenus.libs.nashorn")
-        relocate("net.kyori", "com.extendedclip.deluxemenus.libs.adventure")
         archiveFileName.set("DeluxeMenus-${rootProject.version}.jar")
     }
     java {
         sourceCompatibility = JavaVersion.VERSION_11
-        targetCompatibility = JavaVersion.VERSION_11
+        targetCompatibility = JavaVersion.VERSION_21
     }
 
     processResources {
