@@ -31,10 +31,10 @@ public enum RequirementType {
       Arrays.asList("amount", "placeholder")),
   HAS_EXP(Arrays.asList("has exp", "hasexp", "exp"),
       "Checks if a player has enough exp",
-      Arrays.asList("amount")),
+          List.of("amount")),
   DOES_NOT_HAVE_EXP(Arrays.asList("!has exp", "!hasexp", "!exp"),
       "Checks if a player has enough exp",
-      Arrays.asList("amount")),
+          List.of("amount")),
   HAS_PERMISSION(Arrays.asList("has permission", "has perm", "haspermission", "hasperm", "perm"),
       "Checks if a player has a specific permission", Collections.singletonList("permission")),
   DOES_NOT_HAVE_PERMISSION(
@@ -82,10 +82,10 @@ public enum RequirementType {
   REGEX_DOES_NOT_MATCH(Arrays.asList("!regex matches", "!regex"),
       "Checks if a placeholder parsed string does not match a regex pattern",
       Arrays.asList("input", "regex")),
-  STRING_LENGTH(Arrays.asList("string length"),
+  STRING_LENGTH(List.of("string length"),
   "Checks if the given string's length is between the provided minimum and (optionally) maximum.",
           Arrays.asList("input", "min", "max")),
-  IS_OBJECT(Arrays.asList("is object"),
+  IS_OBJECT(List.of("is object"),
           "Checks if the given string can be parsed as a given Java object.",
           Arrays.asList("input", "object"));
 
