@@ -20,7 +20,6 @@ import com.extendedclip.deluxemenus.utils.Messages;
 import com.extendedclip.deluxemenus.utils.VersionHelper;
 import com.google.common.io.ByteArrayDataOutput;
 import com.google.common.io.ByteStreams;
-import net.kyori.adventure.platform.bukkit.BukkitAudiences;
 import net.kyori.adventure.text.Component;
 import org.bukkit.Bukkit;
 import org.bukkit.Material;
@@ -31,11 +30,7 @@ import org.bukkit.inventory.ItemStack;
 import org.bukkit.plugin.java.JavaPlugin;
 import org.jetbrains.annotations.NotNull;
 
-import java.util.Arrays;
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Objects;
-import java.util.Optional;
+import java.util.*;
 import java.util.function.Function;
 import java.util.logging.Level;
 import java.util.stream.Collectors;
@@ -48,8 +43,6 @@ public class DeluxeMenus extends JavaPlugin {
 
     private PersistentMetaHandler persistentMetaHandler;
     private MenuItemMarker menuItemMarker;
-
-    private BukkitAudiences audiences;
 
     private VaultHook vaultHook;
 
@@ -82,8 +75,6 @@ public class DeluxeMenus extends JavaPlugin {
         this.menuItemMarker = new MenuItemMarker(this);
         new DupeFixer(this, this.menuItemMarker).register();
 
-        this.audiences = BukkitAudiences.create(this);
-
         hookIntoVault();
         setUpItemHooks();
 
@@ -110,11 +101,6 @@ public class DeluxeMenus extends JavaPlugin {
         Bukkit.getMessenger().unregisterOutgoingPluginChannel(this, "BungeeCord");
 
         Bukkit.getScheduler().cancelTasks(this);
-
-        if (this.audiences != null) {
-            this.audiences.close();
-            this.audiences = null;
-        }
 
         Menu.unloadForShutdown(this);
 
@@ -161,11 +147,11 @@ public class DeluxeMenus extends JavaPlugin {
     }
 
     public void sms(CommandSender s, Component msg) {
-        audiences().sender(s).sendMessage(msg);
+        s.sendMessage(msg);
     }
 
     public void sms(CommandSender s, Messages msg) {
-        audiences().sender(s).sendMessage(msg.message());
+        s.sendMessage(msg.message());
     }
 
     public void debug(@NotNull final DebugLevel messageDebugLevel, @NotNull final Level level, @NotNull final String... messages) {
@@ -192,13 +178,6 @@ public class DeluxeMenus extends JavaPlugin {
 
     public PersistentMetaHandler getPersistentMetaHandler() {
         return persistentMetaHandler;
-    }
-
-    public BukkitAudiences audiences() {
-        if (this.audiences == null) {
-            throw new IllegalStateException("Tried to access Adventure when the plugin was disabled!");
-        }
-        return this.audiences;
     }
 
     public void clearCaches() {
